requirements.txt
streamlit == 1.31.1
pandas == 2.1.4
streamlit-authenticator==0.3.1
streamlit-chat==0.1.1
streamlit-folium==0.18.0
streamlit-webrtc==0.47.1
mysql-connector-python==8.3.0
snowflake-connector-python==3.7.0
snowflake-snowpark-python==1.12.1
tokenizers==0.13.3
toml==0.10.2
firebase-admin==6.5.0
python-dotenv==1.0.0
geojson==3.0.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
numpy==1.23.5
matplotlib==3.8.0